feat(week-6): formatted contracts

+++ src/week_6/contracts/Hackathon.sol
@@ -4,25 +4,24 @@ pragma solidity ^0.8.4;
 contract Hackathon {
     struct Project {
         string title;
-        uint[] ratings;
+        uint256[] ratings;
     }
-    
+
     Project[] projects;
 
-    
-    function findWinner() external view returns(Project memory){
+    function findWinner() external view returns (Project memory) {
         Project memory winner;
-        uint max = 0;
+        uint256 max = 0;
 
-        for (uint i =0;i<projects.length;i++){
+        for (uint256 i = 0; i < projects.length; i++) {
             Project memory p = projects[i];
             uint256 avg = 0;
 
-            for (uint256 r = 0; r < p.ratings.length; r++){
+            for (uint256 r = 0; r < p.ratings.length; r++) {
                 avg += p.ratings[r];
             }
 
-            if(avg / p.ratings.length > max){
+            if (avg / p.ratings.length > max) {
                 winner = p;
                 max = avg / p.ratings.length;
             }
@@ -35,7 +34,7 @@ contract Hackathon {
         projects.push(Project(_title, new uint[](0)));
     }
 
-    function rate(uint _idx, uint _rating) external {
+    function rate(uint256 _idx, uint256 _rating) external {
         projects[_idx].ratings.push(_rating);
     }
 }
