+++ src/week_6/contracts/PartySplit.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.4;
 
 contract Party {
     uint256 public amount;
-    mapping (address => bool) public participants;
+    mapping(address => bool) public participants;
     address[] participantsList;
 
     constructor(uint256 _amount) {
@@ -18,18 +18,15 @@ contract Party {
         participantsList.push(msg.sender);
     }
 
-    function payBill(address _venue, uint _billCost) external {
-
+    function payBill(address _venue, uint256 _billCost) external {
         (bool ok,) = _venue.call{value: _billCost}("");
         require(ok);
 
         uint256 refund = address(this).balance / participantsList.length;
 
-        for (uint i = 0; i< participantsList.length; i++){
+        for (uint256 i = 0; i < participantsList.length; i++) {
             (ok,) = participantsList[i].call{value: refund}("");
             require(ok);
         }
     }
-
-
 }
